[rename] #9 카멜케이스로 변경
