+++ app/src/main/java/com/sopt/now/data/dto/response/ResponseFriendDto.kt
@@ -10,13 +10,13 @@ data class ResponseFriendDto(
     @SerialName("page")
     val page: Int,
     @SerialName("per_page")
-    val per_page: Int,
+    val perPage: Int,
     @SerialName("support")
     val support: Support,
     @SerialName("total")
     val total: Int,
     @SerialName("total_pages")
-    val total_pages: Int
+    val totalPages: Int
 ) {
     @Serializable
     data class Data(
@@ -25,11 +25,11 @@ data class ResponseFriendDto(
         @SerialName("email")
         val email: String,
         @SerialName("first_name")
-        val first_name: String,
+        val firstName: String,
         @SerialName("id")
         val id: Int,
         @SerialName("last_name")
-        val last_name: String
+        val lastName: String
     )
 
     @Serializable
