+++ app/src/main/java/com/sopt/now/ui/home/HomeFragment.kt
@@ -6,10 +6,9 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
-import androidx.lifecycle.MutableLiveData
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.sopt.now.data.Friend
-import com.sopt.now.data.Key.USER_ID
+import com.sopt.now.data.KeyStorage.USER_ID
 import com.sopt.now.data.ServicePool
 import com.sopt.now.data.dto.response.ResponseFriendDto
 import com.sopt.now.data.dto.response.ResponseUserDto
@@ -68,7 +67,7 @@ class HomeFragment() : Fragment() {
 
                     friend?.forEach { friendData ->
                         mockFriendList.add(
-                            Friend(friendData.avatar, friendData.first_name, friendData.email)
+                            Friend(friendData.avatar, friendData.firstName, friendData.email)
                         )
                     }
