+++ app/src/main/java/com/sopt/now/ui/home/UserViewHolder.kt
@@ -1,7 +1,6 @@
 package com.sopt.now.ui.home
 
 import androidx.recyclerview.widget.RecyclerView
-import com.sopt.now.data.User
 import com.sopt.now.data.dto.response.ResponseUserDto
 import com.sopt.now.databinding.ItemUserBinding
